Rename primitives.h to libpirate.h
