+++ libpirate/bench/primitives_bench_lat.c
@@ -25,7 +25,7 @@
 #include <time.h>
 #include <unistd.h>
 
-#include "primitives.h"
+#include "libpirate.h"
 
 static int set_buffer_size(int gd, int size) {
   size = 8 * size;
