+++ libpirate/libpirate.h
