+++ libpirate/test/channel_test.cpp
@@ -15,7 +15,7 @@
 
 #include <cstring>
 #include <stdlib.h>
-#include "primitives.h"
+#include "libpirate.h"
 #include "channel_test.hpp"
 
 namespace GAPS
