+++ libpirate/udp_socket.h
@@ -16,7 +16,7 @@
 #ifndef __PIRATE_CHANNEL_UDP_SOCKET_H
 #define __PIRATE_CHANNEL_UDP_SOCKET_H
 
-#include "primitives.h"
+#include "libpirate.h"
 
 typedef struct {
     int sock;
